Change unchanged file detection to digest based.
mtime is updated when a file is modified or created
ctime is updated when a file is modified, created, or
(commonly) ownership/permissions are changed.
If the expectation is that the skel file creates this
and then the file is updated to have the ring code,
this method won't discover that.  Usually, you'll want
to compare it to the skel file using either md5 or diff.
More likely, you'll want to keep each version of the
skel file in a digest in this file as an unedited list
so that any time it changes, you'll be able to not only
continue to see that older ones haven't been edited
still but also what version of non-editness they are.

This commit does that (untested thus far since I'm editing in github directly but it should work).